refactoring and unit tests
- extracted some code (`InMemoryFileSystem` and related) to separate file
- basic unit tests to ensure TypeScript libraries are exposed/hidden when needed
